윤정 1108 16:15

+++ Dsup/src/main/java/com/dsup/dbmanagement/UserTbspcTbVO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class UserTbspcTbVO {
 	private String tablespaceName;
 	private String userId;
-	private float volumn;
+	private double volumn;
 }
